♻️ Refactor: DocDetail 페이지 컴포넌트 및 데이터 구조 분리
